+++ R/hw04pgnmergen613-package.R
@@ -1,0 +1,10 @@
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+#' @importFrom tibble tibble tribble
+#' @importFrom ggplot2 ggplot aes geom_line theme labs theme_minimal element_text
+## usethis namespace: end
+NULL