+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(hw04pgnmergen613)
+
+test_check("hw04pgnmergen613")