+++ NEWS.md
@@ -1,3 +1,15 @@
-# hw04pgnmergen613 0.0.0.9000
+# hw04pgnmergen 0.0.1.0002
+
+* Added test.R to tests functions
+
+# hw04pgnmergen 0.0.1.0001
+
+* Added function plotter. See ?plotter for details.
+
+# hw04pgnmergen 0.0.1.0000
+
+* Added function myseq_n. ?fun_name directory added with examples.
+
+# hw04pgnmergen 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.